Create a submodule nn; Move all the model related stuff, EmbeddingColumn, ClassificationOutputColumn, SegmentationOutputColumn, InstancesColumn to meerkat.nn
