+++ meerkat/__init__.py
@@ -9,28 +9,23 @@ from meerkat.logging.utils import (
 
 initialize_logging()
 
+from meerkat import nn
 from meerkat.cells.abstract import AbstractCell
 from meerkat.cells.imagepath import ImagePath
 from meerkat.cells.spacy import LazySpacyCell, SpacyCell
 from meerkat.cells.volume import MedicalVolumeCell
 from meerkat.columns.abstract import AbstractColumn
 from meerkat.columns.cell_column import CellColumn
-from meerkat.columns.embedding_column import EmbeddingColumn
 from meerkat.columns.image_column import ImageColumn
-from meerkat.columns.instances_column import InstancesColumn
 from meerkat.columns.lambda_column import LambdaCell, LambdaColumn
 from meerkat.columns.list_column import ListColumn
 from meerkat.columns.numpy_column import NumpyArrayColumn
 from meerkat.columns.pandas_column import PandasSeriesColumn
-from meerkat.columns.prediction_column import ClassificationOutputColumn
-from meerkat.columns.segmentation_column import SegmentationOutputColumn
 from meerkat.columns.spacy_column import SpacyColumn
 from meerkat.columns.tensor_column import TensorColumn
 from meerkat.columns.video_column import VideoColumn
 from meerkat.columns.volume_column import MedicalVolumeColumn
 from meerkat.datapanel import DataPanel
-from meerkat.model.huggingfacemodel import HuggingfaceModel
-from meerkat.model.tensormodel import TensorModel
 from meerkat.ops.concat import concat
 from meerkat.ops.merge import merge
 from meerkat.provenance import provenance
