+++ meerkat/nn/__init__.py
@@ -1,0 +1,9 @@
+"""Import nn module classes"""
+# flake8: noqa
+
+from meerkat.nn.embedding_column import EmbeddingColumn
+from meerkat.nn.huggingfacemodel import HuggingfaceModel
+from meerkat.nn.instances_column import InstancesColumn
+from meerkat.nn.prediction_column import ClassificationOutputColumn
+from meerkat.nn.segmentation_column import SegmentationOutputColumn
+from meerkat.nn.tensormodel import TensorModel
