+++ meerkat/nn/huggingfacemodel.py
@@ -1,27 +1,29 @@
 from __future__ import annotations
 
 import itertools
-from typing import Dict, List
+from typing import Dict, List, Optional
 
 import cytoolz as tz
 import torch
 from tqdm import tqdm
 
-from meerkat import DataPanel
-from meerkat.columns.embedding_column import EmbeddingColumn
-from meerkat.columns.instances_column import InstancesColumn
-from meerkat.columns.prediction_column import ClassificationOutputColumn
-from meerkat.columns.segmentation_column import SegmentationOutputColumn
-from meerkat.columns.tensor_column import TensorColumn
-from meerkat.model.activation import ActivationOp
-from meerkat.model.model import Model
+from meerkat.columns.list_column import ListColumn
+from meerkat.datapanel import DataPanel
+from meerkat.nn.activation import ActivationOp
+from meerkat.nn.embedding_column import EmbeddingColumn
+from meerkat.nn.model import Model
+from meerkat.nn.prediction_column import ClassificationOutputColumn
+from meerkat.tools.lazy_loader import LazyLoader
+
+AutoTokenizer = LazyLoader("transformers.AutoTokenizer")
 
 
-class TensorModel(Model):
+class HuggingfaceModel(Model):
     def __init__(
         self,
-        # identifier: str,
+        identifier: str,
         model,
+        tokenizer: Optional[AutoTokenizer] = None,
         is_classifier: bool = None,
         task: str = None,
         device: str = None,
@@ -29,9 +31,10 @@ class TensorModel(Model):
 
         if model is None:
             raise ValueError(
-                f"A PyTorch model is required with {self.__class__.__name__}."
+                f"A HuggingFace model is required with {self.__class__.__name__}."
             )
-        super(TensorModel, self).__init__(
+
+        super(HuggingfaceModel, self).__init__(
             # identifier=identifier,
             model=model,
             is_classifier=is_classifier,
@@ -39,41 +42,62 @@ class TensorModel(Model):
             device=device,
         )
 
+        self.tokenizer = tokenizer
+        if tokenizer is None:
+            # Load the tokenizer
+            self.tokenizer = AutoTokenizer.from_pretrained(self.identifier)
+
         # Move the model to device
         self.to(self.device)
 
     def forward(self, input_batch: Dict) -> Dict:
-        # Run the model on the input_batch
-        with torch.no_grad():
-            outputs = self.model(input_batch)
 
         if self.is_classifier:
+            # Run the model on the input_batch
+            with torch.no_grad():
+                outputs = self.model(**input_batch)
+
             # probs and preds can be handled at ClassificationOutputColumn
             # TODO(Priya): See if there is any case where these are to be returned
-            output_dict = {"logits": outputs.to("cpu")}
+            # Logits are present at the 0th index
+            output_dict = {"logits": outputs[0].to("cpu")}
 
-        elif self.task == "semantic_segmentation":
-            # Output is a dict with key 'out'
-            output_dict = {"logits": outputs["out"].to("cpu")}
-
-        elif self.task == "timeseries":
-            output_dict = {"preds": outputs.to("cpu")}
-
-        elif self.task == "instance_segmentation":
-            output_dict = {
-                "preds": [output["instances"].to("cpu") for output in outputs]
-            }
+        else:
+            # TODO (Priya): Support for only summarization right now.
+            with torch.no_grad():
+                summary_token_ids = self.model.generate(**input_batch)
+                summaries = [
+                    self.tokenizer.decode(
+                        token_id_list,
+                        skip_special_tokens=True,
+                        clean_up_tokenization_spaces=False,
+                    )
+                    for token_id_list in summary_token_ids
+                ]
+                output_dict = {"preds": summaries}
 
         return output_dict
 
+    def encode_batch(self, batch: DataPanel, columns: List[str], **kwargs):
+        # TODO(karan): Automatically writing this encoder for a variety of tasks
+        return self.tokenizer(
+            *[list(batch[key]) for key in columns],
+            truncation=True,
+            padding=True,
+            **kwargs,
+        )
+
     def process_batch(self, batch: DataPanel, input_columns: List[str]):
 
-        # Convert the batch to torch.Tensor and move to device
-        if self.task == "instance_segmentation":
-            input_batch = batch[input_columns[0]].data
-        else:
-            input_batch = batch[input_columns[0]].data.to(self.device)
+        # Tokenize the batch
+        input_batch = self.encode_batch(batch=batch, columns=input_columns)
 
+        # Convert the batch to torch.Tensor
+        input_batch = tz.valmap(
+            lambda v: torch.tensor(v).to(device=self.device), input_batch
+        )
+
+        # Return the converted batch
         return input_batch
 
     def predict_batch(
@@ -117,7 +141,7 @@ class TensorModel(Model):
 
             # Forward pass
             with torch.no_grad():
-                self.model(input_batch)
+                self.model(**input_batch)
 
             # Get activations for the batch
             batch_activation = {
@@ -135,6 +159,7 @@ class TensorModel(Model):
         dataset.add_column(f"activation ({target_module})", activation_col)
         return activation_col
 
+    # TODO(Priya): Need to test on NLP model
     def classification(
         self,
         dataset: DataPanel,
@@ -146,12 +171,12 @@ class TensorModel(Model):
         threshold=0.5,
     ) -> DataPanel:
 
-        # Handles outputs for classification tasks
-
         predictions = self.predict_batch(dataset, input_columns, batch_size)
         predictions = tz.merge_with(lambda v: torch.cat(v).to("cpu"), *predictions)
 
         logits = predictions["logits"]
+        # Store in correct column type
+        # TODO(Priya): Better way for feeding classifier input
         output_col = ClassificationOutputColumn(
             logits=logits,
             num_classes=num_classes,
@@ -168,78 +193,28 @@ class TensorModel(Model):
             }
         )
         # TODO(Priya): Use this instead after append bug is resolved
-        # dataset = dataset.append(classifier_dp, axis=1)
+        # dataset = dataset.append(output_dp, axis=1)
         dataset.add_column("logits", output_col)
         dataset.add_column("probs", output_col.probabilities())
         dataset.add_column("preds", output_col.predictions())
 
         return output_dp
 
-    def semantic_segmentation(
-        self,
-        dataset: DataPanel,
-        input_columns: List[str],
-        batch_size: int = 32,
-        num_classes: int = None,
-    ) -> DataPanel:
-
-        # Handles outputs for semantic_segmentation tasks
-
-        predictions = self.predict_batch(dataset, input_columns, batch_size)
-        predictions = tz.merge_with(lambda v: torch.cat(v).to("cpu"), *predictions)
-
-        logits = predictions["logits"]
-        output_col = SegmentationOutputColumn(logits=logits, num_classes=num_classes)
-
-        output_dp = DataPanel(
-            {
-                "logits": output_col,
-                "probs": SegmentationOutputColumn(output_col.probabilities().data),
-                "preds": SegmentationOutputColumn(output_col.predictions().data),
-            }
-        )
-        # TODO(Priya): Uncomment after append bug is resolved
-        # dataset = dataset.append(classifier_dp, axis=1)
-        dataset.add_column("logits", output_col)
-        dataset.add_column("probs", output_col.probabilities())
-        dataset.add_column("preds", output_col.predictions())
-
-        return output_dp
-
-    def timeseries(
+    def summarization(
         self, dataset: DataPanel, input_columns: List[str], batch_size: int = 32
     ) -> DataPanel:
-
-        # Handles outputs for timeseries
-
-        predictions = self.predict_batch(dataset, input_columns, batch_size)
-        predictions = tz.merge_with(lambda v: torch.cat(v).to("cpu"), *predictions)
-
-        output_col = TensorColumn(predictions["preds"])
-        output_dp = DataPanel({"preds": output_col})
-
-        # TODO(Priya): Uncomment after append bug is resolved
-        # dataset = dataset.append(classifier_dp, axis=1)
-        dataset.add_column("preds", output_col)
-
-        return output_dp
-
-    def instance_segmentation(
-        self, dataset: DataPanel, input_columns: List[str], batch_size: int = 32
-    ) -> DataPanel:
-
-        # Handles outputs for instance segmentation
 
         predictions = self.predict_batch(dataset, input_columns, batch_size)
         predictions = tz.merge_with(
-            lambda v: list(itertools.chain.from_iterable(v)), *predictions
+            lambda x: list(itertools.chain.from_iterable(x)), *predictions
         )
 
-        output_col = InstancesColumn(predictions["preds"])
+        # Store in correct column type
+        output_col = ListColumn(predictions["preds"])
         output_dp = DataPanel({"preds": output_col})
 
         # TODO(Priya): Uncomment after append bug is resolved
-        # dataset = dataset.append(classifier_dp, axis=1)
+        # dataset = dataset.append(output_dp, axis=1)
         dataset.add_column("preds", output_col)
 
         return output_dp
@@ -254,7 +229,7 @@ class TensorModel(Model):
         one_hot: bool = None,
         threshold=0.5,
     ):
-
+        # TODO(Priya): The separate functions can be merged later
         if self.is_classifier:
             return self.classification(
                 dataset,
@@ -265,13 +240,19 @@ class TensorModel(Model):
                 one_hot,
                 threshold,
             )
-        elif self.task == "semantic_segmentation":
-            return self.semantic_segmentation(
-                dataset, input_columns, batch_size, num_classes
-            )
-        elif self.task == "timeseries":
-            return self.timeseries(dataset, input_columns, batch_size)
-        elif self.task == "instance_segmentation":
-            return self.instance_segmentation(dataset, input_columns, batch_size)
+        elif self.task == "summarization":
+            return self.summarization(dataset, input_columns, batch_size)
         else:
             raise NotImplementedError
+
+    @classmethod
+    def _state_keys(cls) -> set:
+        """List of attributes that describe the state of the object."""
+        return super()._state_keys() | {
+            "identifier",
+            "model",
+            "tokenizer",
+            "is_classifier",
+            "task",
+            "device",
+        }
