+++ tests/meerkat/nn/test_segmentation_column.py
@@ -1,8 +1,8 @@
 import pytest
 import torch
 
-from meerkat.columns.segmentation_column import SegmentationOutputColumn
 from meerkat.columns.tensor_column import TensorColumn
+from meerkat.nn.segmentation_column import SegmentationOutputColumn
 
 logits_multi = torch.tensor(
     [
