+++ tests/meerkat/pipelines/test_entitydatapanel.py
@@ -5,7 +5,8 @@ import numpy as np
 import pytest
 import torch
 
-from meerkat import DataPanel, EmbeddingColumn, ListColumn, NumpyArrayColumn
+from meerkat import DataPanel, ListColumn, NumpyArrayColumn
+from meerkat.nn import EmbeddingColumn
 from meerkat.pipelines.entitydatapanel import EntityDataPanel
